started a base audiovideo class
