+++ src/pynaviz/audiovideo/video_handling.py
@@ -10,35 +10,9 @@ import numpy as np
 
 # from line_profiler import profile
 from numpy.typing import NDArray
+from .base_audiovideo import BaseAudioVideo
 
-
-def ts_to_index(ts: float, time: NDArray) -> int:
-    """
-    Return the index of the frame whose experimental time is just before (or equal to) `ts`.
-
-    Parameters
-    ----------
-    ts : float
-        Experimental timestamp to match.
-    time : NDArray
-        Array of experimental timestamps, assumed sorted in ascending order,
-        with one entry per frame.
-
-    Returns
-    -------
-    idx : int
-        Index of the frame with time <= `ts`. Clipped to [0, len(time) - 1].
-
-    Notes
-    -----
-    - If `ts` is smaller than all values in `time`, returns 0.
-    - If `ts` is greater than all values in `time`, returns `len(time) - 1`.
-    """
-    idx = np.searchsorted(time, ts, side="right") - 1
-    return np.clip(idx, 0, len(time) - 1)
-
-
-class VideoHandler:
+class VideoHandler(BaseAudioVideo):
     """Class for getting video frames."""
 
     _thread_local = threading.local()
@@ -50,13 +24,11 @@ class VideoHandler:
         time: Optional[NDArray] = None,
         return_frame_array: bool = True,
     ) -> None:
-        self._thread_local.get_from_index = False
-        self.video_path = pathlib.Path(video_path)
-        self.container = av.open(video_path)
+        super().__init__(video_path)
         self.stream = self.container.streams.video[stream_index]
         self.stream_index = stream_index
         self.return_frame_array = return_frame_array
-        self._running = True
+
 
         # default to linspace
         # TODO : what if number of frames is 0.
@@ -78,7 +50,7 @@ class VideoHandler:
         # initialize current frame
         self.current_frame: Optional[av.VideoFrame] = None
 
-        if self.video_path.suffix == ".mkv":
+        if self.file_path.suffix == ".mkv":
             # mkv time is rounded to 3 digits, at least in the example video
             # generated by tests/generate_numbered_video.py
             self.round_fn = lambda x: np.round(x, 3)
@@ -91,7 +63,6 @@ class VideoHandler:
         self.key_mask = None
 
         self._i = 0  # write position
-        self._lock = threading.Lock()
         if self.stream.frames and self.stream.frames > 0:
             self._index_thread = threading.Thread(target=self._build_index_fixed_size, daemon=True)
         else:
@@ -99,10 +70,33 @@ class VideoHandler:
 
         self._index_ready = threading.Event()
         self._index_thread.start()
-        self._keyframe_pts = []
-        self._pts_keyframe_ready = threading.Event()
-        self._keyframe_thread = threading.Thread(target=self._extract_keyframes_pts, daemon=True)
-        self._keyframe_thread.start()
+
+
+    @staticmethod
+    def ts_to_index(ts: float, time: NDArray) -> int:
+        """
+        Return the index of the frame whose experimental time is just before (or equal to) `ts`.
+
+        Parameters
+        ----------
+        ts : float
+            Experimental timestamp to match.
+        time : NDArray
+            Array of experimental timestamps, assumed sorted in ascending order,
+            with one entry per frame.
+
+        Returns
+        -------
+        idx : int
+            Index of the frame with time <= `ts`. Clipped to [0, len(time) - 1].
+
+        Notes
+        -----
+        - If `ts` is smaller than all values in `time`, returns 0.
+        - If `ts` is greater than all values in `time`, returns `len(time) - 1`.
+        """
+        idx = np.searchsorted(time, ts, side="right") - 1
+        return np.clip(idx, 0, len(time) - 1)
 
     def extract_keyframe_times_and_points(
         self, video_path: str | pathlib.Path, stream_index: int = 0, first_only=False
@@ -176,7 +170,7 @@ class VideoHandler:
 
     def _extract_keyframes_pts(self):
         try:
-            with av.open(self.video_path) as container:
+            with av.open(self.file_path) as container:
                 stream = container.streams.video[0]
                 for packet in container.demux(stream):
                     if not self._running:
@@ -192,7 +186,7 @@ class VideoHandler:
 
     def _build_index_fixed_size(self):
         try:
-            with av.open(self.video_path) as container:
+            with av.open(self.file_path) as container:
                 stream = container.streams.video[self.stream_index]
                 n_frames = stream.frames
 
@@ -218,7 +212,7 @@ class VideoHandler:
 
     def _build_index_dynamic(self):
         try:
-            with av.open(self.video_path) as container:
+            with av.open(self.file_path) as container:
                 if not self._running:
                     return
                 stream = container.streams.video[self.stream_index]
@@ -240,25 +234,25 @@ class VideoHandler:
         finally:
             self._index_ready.set()
 
-    def _need_seek_call(self, current_frame_pts, target_frame_pts):
-        with self._lock:
-            # return if empty list or empty array or not enough frmae
-            if len(self._keyframe_pts) == 0 or self._keyframe_pts[-1] < target_frame_pts:
-                return True
-
-        # roll back the stream if video is scrolled backwards
-        if current_frame_pts > target_frame_pts:
-            return True
-
-        # find the closest keyframe pts before a given frame
-        idx = np.searchsorted(self._keyframe_pts, target_frame_pts, side="right")
-        closest_keyframe_pts = self._keyframe_pts[max(0, idx - 1)]
-
-        # if target_frame_pts is larger than current (and if code
-        # arrives here, it is, see second return statement),
-        # then seek forward if there is a future keyframe closest
-        # to the target.
-        return closest_keyframe_pts > current_frame_pts
+    # def _need_seek_call(self, current_frame_pts, target_frame_pts):
+    #     with self._lock:
+    #         # return if empty list or empty array or not enough frmae
+    #         if len(self._keyframe_pts) == 0 or self._keyframe_pts[-1] < target_frame_pts:
+    #             return True
+    #
+    #     # roll back the stream if video is scrolled backwards
+    #     if current_frame_pts > target_frame_pts:
+    #         return True
+    #
+    #     # find the closest keyframe pts before a given frame
+    #     idx = np.searchsorted(self._keyframe_pts, target_frame_pts, side="right")
+    #     closest_keyframe_pts = self._keyframe_pts[max(0, idx - 1)]
+    #
+    #     # if target_frame_pts is larger than current (and if code
+    #     # arrives here, it is, see second return statement),
+    #     # then seek forward if there is a future keyframe closest
+    #     # to the target.
+    #     return closest_keyframe_pts > current_frame_pts
 
     def _get_frame_idx(self, pts: int) -> int:
         """
@@ -399,7 +393,7 @@ class VideoHandler:
 
     def get(self, ts: float) -> av.VideoFrame | NDArray:
         if not getattr(self._thread_local, "get_from_index", False):
-            idx = ts_to_index(ts, self.time)
+            idx = self.ts_to_index(ts, self.time)
         else:
             idx = ts
 
@@ -519,21 +513,21 @@ class VideoHandler:
     def t(self):
         return self.time
 
-    def close(self):
-        """Close the video stream."""
-        self._running = False
-        if self._index_thread.is_alive():
-            self._index_thread.join(timeout=1)  # Be conservative, don’t block forever
-        if self._keyframe_thread.is_alive():
-            self._keyframe_thread.join(timeout=1)
-        try:
-            self.container.close()
-        except Exception:
-            print("VideoHandler failed to close the video stream.")
-        finally:
-            # dropping refs to fully close av.InputContainer
-            self.container = None
-            self.stream = None
+    # def close(self):
+    #     """Close the video stream."""
+    #     self._running = False
+    #     if self._index_thread.is_alive():
+    #         self._index_thread.join(timeout=1)  # Be conservative, don’t block forever
+    #     if self._keyframe_thread.is_alive():
+    #         self._keyframe_thread.join(timeout=1)
+    #     try:
+    #         self.container.close()
+    #     except Exception:
+    #         print("VideoHandler failed to close the video stream.")
+    #     finally:
+    #         # dropping refs to fully close av.InputContainer
+    #         self.container = None
+    #         self.stream = None
 
     def _wait_for_index(self, timeout=2.0):
         """Wait up to timeout.
@@ -546,9 +540,9 @@ class VideoHandler:
 
     def get_slice(self, start: float, end: float = None):
         # TODO check start and end are sorted
-        start = ts_to_index(start, self.time)
+        start = self.ts_to_index(start, self.time)
         if end:
-            end = ts_to_index(end, self.time)
+            end = self.ts_to_index(end, self.time)
             return slice(start, end)
         else:
             return slice(start, start + 1)
@@ -725,11 +719,3 @@ class VideoHandler:
 
     def __len__(self):
         return self.shape[0]
-
-    # context protocol
-    # (with VideoHandler(path) as video: ensure closing)
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.close()
